+++ amundsen_application/static/js/components/SVGIcons/AlertIcon.tsx
