+++ amundsen_application/static/js/components/SVGIcons/LeftIcon.tsx
