+++ amundsen_application/static/js/components/SVGIcons/RightIcon.tsx
