+++ amundsen_application/static/js/components/SVGIcons/UpIcon.tsx
