+++ amundsen_application/static/js/components/SVGIcons/constants.ts
