+++ amundsen_application/static/js/components/SVGIcons/index.tsx
