+++ amundsen_application/static/js/components/SVGIcons/types.ts
