+++ amundsen_application/static/js/components/common/Table/index.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 
 import ShimmeringResourceLoader from '../ShimmeringResourceLoader';
-import { UpIcon, DownIcon } from '../SVGIcons';
+import { UpIcon, DownIcon } from '../../SVGIcons';
 
 import './styles.scss';
